Solved the bottom navigation fault
